Implement `TakeHandle<_>` for `Tracer`
